requirements.txt
Flask==3.0.3
pypdf==4.3.1
pdf2image==1.17.0
Pillow==10.4.0
Flask-WTF==1.2.1
Werkzeug==3.0.3
gunicorn==22.0.0
